requirements.txt
langchain
langchain-community
huggingface_hub
requests
python-dotenv
